feat(types): Exposed catalog well-known items CRDs
