+++ .changeset/poor-tools-trade.md
@@ -1,0 +1,5 @@
+---
+"@mia-platform/console-types": patch
+---
+
+Exposed catalog well-known items CRDs
