--- packages/console-types/src/types/catalog/custom-resource-definitions/index.ts
@@ -1,28 +1,0 @@
-/**
- * Copyright 2025 Mia srl
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { CatalogCRDManifest } from '../well-known-items'
-
-/**
- * This type is meant only as a way to ensure correctness of public CRDs.
- * It should not be exposed nor used elsewhere.
- */
-export type PublicCatalogCRD = CatalogCRDManifest & {
-  tenantId: 'mia-platform',
-  visibility: { public: true },
-}
