--- packages/console-types/src/types/catalog/custom-resource-definitions/infrastructure-resource.crd.ts
@@ -1,40 +1,0 @@
-/**
- * Copyright 2025 Mia srl
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { catalogInfrastructureResource } from '../well-known-items'
-import { PublicCatalogCRD } from '.'
-
-const infrastructureResourceCRD: PublicCatalogCRD = {
-  name: 'custom-resource',
-  itemId: 'custom-resource',
-  description: 'Custom Workload Resource Definition',
-  type: 'custom-resource-definition',
-  tenantId: 'mia-platform',
-  isVersioningSupported: true,
-  visibility: { public: true },
-  resources: {
-    name: 'custom-resource',
-    validation: {
-      jsonSchema: {
-        ...catalogInfrastructureResource.resourcesSchema,
-      },
-    },
-  },
-}
-
-export default infrastructureResourceCRD
