+++ packages/console-types/src/types/catalog/well-known-items/infrastructure-resource/index.ts
@@ -23,6 +23,7 @@ import { catalogNameSchema } from '../commons'
 import { CatalogItemManifest } from '../../item-manifest'
 import { CatalogItem } from '../../item'
 import { CatalogVersionedItem } from '../../versioned-item'
+import { PublicCatalogCRD, CatalogCRDManifest } from '../custom-resource-definition'
 
 const type = 'custom-resource'
 
@@ -117,6 +118,24 @@ const resourcesSchema = {
   type: 'object',
 } as const satisfies JSONSchema
 
+const crd: CatalogCRDManifest = {
+  name: 'custom-resource',
+  itemId: 'custom-resource',
+  description: 'Custom Workload Resource Definition',
+  type: 'custom-resource-definition',
+  tenantId: 'mia-platform',
+  isVersioningSupported: true,
+  visibility: { public: true },
+  resources: {
+    name: type,
+    validation: {
+      jsonSchema: {
+        ...resourcesSchema,
+      },
+    },
+  },
+} satisfies PublicCatalogCRD
+
 export type CatalogInfrastructureResourceResources = FromSchema<typeof resourcesSchema>
 export type CatalogInfrastructureResourceItem = CatalogItem<typeof type, CatalogInfrastructureResourceResources>
 export type CatalogInfrastructureResourceVersionedItem = CatalogVersionedItem<
@@ -128,4 +147,4 @@ export type CatalogInfrastructureResourceManifest = CatalogItemManifest<
   CatalogInfrastructureResourceResources
 >
 
-export default { type, resourcesSchema }
+export default { type, resourcesSchema, crd }
